Deduplicate slowapi test stub
